requirements.txt
Flask>=2.3,<3
Werkzeug>=2.3,<3
gunicorn>=21.2,<21.3
psycopg2-binary>=2.9,<3
